requirements.txt
# code dependencies
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=8.0.0
pytest-html>=4.1.0
pytest-metadata>=3.1.0